### 2016-07-13 #58 @sckott
getlandsat package
- [ ] What does this package do? (explain in 50 words or less):

`getlandsat` provides access to Landsat (<https://landsat.usgs.gov/>) 8 metadata and images hosted on AWS S3 at <https://aws.amazon.com/public-data-sets/landsat/> The package only fetches data for users, and does not aid in downstream usage, but additional functionality may be added if deemed necessary.

- [ ] Paste the full DESCRIPTION file inside a code block below:

```
Package: getlandsat
Title: Get Landsat 8 Data
Description: Get Landsat 8 Data from AWS public data sets. Includes
    functions for listing images and fetching them.
Version: 0.0.6.9000
Authors@R: c(person("Scott", "Chamberlain", role = c("aut", "cre"),
    email = "myrmecocystus@gmail.com"))
    readr (>= 0.2.2),
    xml2,
    rappdirs
    covr
```

- [x] URL for the package (the development repository, not a stylized html page):

https://github.com/ropenscilabs/getlandsat

- [x] Who is the target audience?

Anyone that wants to use Landsat images - ecologists, map makers, etc.

- [ ] Are there other R packages that accomplish the same thing? If so, what is different about yours?

None that I know of

- [ ] Do you intend for this package to go on CRAN?  No
    - [ ] The package contains a [`paper.md`](http://joss.theoj.org/about#paper_structure) with a high-level description.
