### 2016-07-13 #58 @noamross
### Editor checks:

- [x] **Fit**: The package meets or criteria [fit](policies.md#fit) and [overlap](policies.md#fit)

---

#### Editor comments

Two things from initial check, not enough to hold up review

- One NOTE: Non-standard file/directory found at top level:  ‘Makefile’
- Code coverage appears only partially set up: there's a badge in the repository but I think the appropriate line is not .travis.yml to run `covr()`

---

Reviewers:
Due date:
