### 2016-07-21 #58 @andeek
---

This package is very well done. I especially liked the caching functionality as the images can be somewhat large and slow to download. The functionality seems very self explanatory except for one thing. I am unsure of when to use `lsat_list()` vs. `lsat_scenes()`. Perhaps a distinction could be added to the vignette. I have included below individual comments based on each portion of the reviewer template.

**A statement of need**

  This is partially true, there is a clear statement in the README of what the package does, but not who the target audience is. Perhaps the section from the summary in the review issue could be added to the README as a more thorough explanation of the package goals and audience.

**Installation instructions:**

  Installation instructions in the README worked perfectly and were very clear.

**Vignette(s)**

  The vignette ran successfully and does demonstrate the major functionality in the paper. It might be helpful to talk a little bit in the vignette about what the `list_scenes` and `list_scene_files` functions are doing and what they return in addition to showing their use.

**Function Documentation:**

  All exported functions are well documented in help. Easy to read and understand.

**Examples**

  I ran into a problem with the following example:

  ```r
  lsat_cache_delete(files = lsat_cache_list()[1])
  ```

  within the function, there is the following line that duplicates a portion of the path.

  ```r
  files <- file.path(lsat_path(), "L8", dat$wrs_path, dat$wrs_row,
        dat$str, files)
  ```

  `dat$wrs_path, dat$wrs_row, dat$str` are all included in the path from `lsat_cache_list()[1]` so that when I run the code I receive the following error:

  ```r
  Error: These files don't exist or can't be found:
     /Users/andeek/Library/Caches/landsat-pds/L8/001/003/LC80010032014272LGN00/L8/001/003/LC80010032014272LGN00/LC80010032014272LGN00_B1.TIF.ovr
  ```

  Other than that, all examples work as expected.

**Community guidelines**

  I found a contributor code of conduct, but not a guide for how to contribute. The DESCRIPTION file is complete with `URL`, `Maintainer` and `BugReports`.

**Installation:**

  No problems with installation.

**Functionality:**

  See the comment on examples for issue with `lsat_cache_delete`. Other than that small problem, the software functions great!
**Automated tests:**

  There were no unit tests for the caching functions, but I'm not sure if that's because it's specific to the user's machine so would be harder to write? All the other functions are tested extensively and pass on my machine.

**Packaging guidelines**

* Package naming - great name!
* Function/variable naming - all functions are names with snake case and avoid name conflicts.
* *README - has almost everything required. The only thing I don't see is the rOpenSci footer image.*
* Code of conduct - included
* Documentation - all the functions are documented very well and there is a vignette as well as top-level documentation.
* *News - this is missing*
* Authorship - looks good
* Package dependencies - looks good
* *Testing - `testthat` is being used, see my above comment on testing the cache functions.*
* Continuous integration - using Travis CI
* Console messages - `cat` is only used in `print.*()` methods and otherwaise `warning` and `message` are used.
* Recommended software scaffolding - using `httr`
* Miscellaneous CRAN gotchas - these all look good
