### 2016-07-28 ropenscilabs/getlandsat#16 @sckott
review: clear up usage of `list_list`/`list_scene_files`/`lsat_scenes`
from review:

> talk a little bit in the vignette about what the `list_scenes` and `list_scene_files` functions are doing and what they return
