### 2016-07-28 #58 @sckott
> One NOTE: Non-standard file/directory found at top level: ‘Makefile’

will rbuildignore that https://github.com/ropenscilabs/getlandsat/issues/13

> Code coverage appears only partially set up: there's a badge in the repository but I think the appropriate line is not .travis.yml to run covr()

i think i stopped doing it cause it was failing for some weird reason, will get it back started

-------

> This package is very well done. I especially liked the caching functionality as the images can be somewhat large and slow to download.

thanks!

> The functionality seems very self explanatory except for one thing. I am unsure of when to use `lsat_list()` vs. `lsat_scenes()` Perhaps a distinction could be added to the vignette.

Thanks, i'll clarify https://github.com/ropenscilabs/getlandsat/issues/14

> there is a clear statement in the README of what the package does, but not who the target audience is. Perhaps the section from the summary in the review issue could be added to the README as a more thorough explanation of the package goals and audience.

thanks, will do  https://github.com/ropenscilabs/getlandsat/issues/15

> vignette .... It might be helpful to talk a little bit in the vignette about what the `list_scenes` and `list_scene_files` functions are doing and what they return in addition to showing their use.

okay, will do https://github.com/ropenscilabs/getlandsat/issues/16

> I ran into a problem with the following example:  `lsat_cache_delete(files = lsat_cache_list()[1])`

I'll fix that and write a test for it, thanks   https://github.com/ropenscilabs/getlandsat/issues/17

> I found a contributor code of conduct, but not a guide for how to contribute.

will add a `CONTRIBUTING.md` file https://github.com/ropenscilabs/getlandsat/issues/18

> There were no unit tests for the caching functions ...

i think I avoided that as they would take a while to run, but i'll try to add some https://github.com/ropenscilabs/getlandsat/issues/19

>  README - has almost everything required. The only thing I don't see is the rOpenSci footer image.

thanks for noting, will add

> News - this is missing

thanks for noting, will add
