### 2016-07-28 ropensci/getlandsat#14 @sckott
closing, merging with #16
