### 2016-08-10 #58 @sckott
@andeek everything has been addressed now

@noamross anything else, or good to go?
