### 2016-08-15 #58 @andeek
Everything above has been addressed and I've edited the comment to reflect that. Thanks @sckott and @noamross!
