### 2016-08-15 #58 @sckott
thanks again @andeek  for the review

@noamross `Date` removed  https://github.com/ropenscilabs/getlandsat/commit/0252fb332a10c045878ae87f2225af43b62232f6
