### 2016-08-15 #58 @noamross
Final checks all pass. Approved!  @sckott you know the drill from here.
